Expanded Sections 1 & 2, Started on section 3
Expanded section 1 to overall be larger/longer and had it include pressure plate to move moving platforms and open tigger doors.

Expanded section 2 past acid lake to include more enemies that become more dense as the section goes along, with the key 2 at the end of the section.

Started creating lower section 3, plans to make it a puzzle centred section with the main focus of using pushable boxes as acid floats.